% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gitbook}
\alias{gitbook}
\title{The GitBook output format with dreal template}
\usage{
gitbook(..., css = NULL, lib_dir = "libs", offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[bookdown]{gitbook}}}

\item{css}{One or more css files to include}

\item{lib_dir}{...
Arguments to be passed to \code{rmarkdown::\link{html_document}()}
(\code{...} not including \code{toc}, and \code{theme}).}

\item{offline}{Set to TRUE to create a bookdown without internet connection}
}
\description{
The GitBook output format with dreal template
}
