% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{html_chapters}
\alias{html_chapters}
\title{Build book chapters into separate HTML files with dreal template}
\usage{
html_chapters(..., offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[bookdown]{html_chapters}}}

\item{offline}{Logical. TRUE if no access to internet}
}
\description{
Build book chapters into separate HTML files with dreal template
}
