% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{html_document2}
\alias{html_document2}
\title{Output formats that allow numbering and cross-referencing figures/tables/equations with dreal template}
\usage{
html_document2(..., offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[bookdown]{html_document2}}}

\item{offline}{Logical. TRUE if no access to internet}
}
\description{
Output formats that allow numbering and cross-referencing figures/tables/equations with dreal template
}
