% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{html_mark}
\alias{html_mark}
\title{Convert to HTML document with dreal template}
\usage{
html_mark(..., offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[rmarkdown]{html_document}}}

\item{offline}{Logical. TRUE if no access to internet}
}
\description{
Convert to HTML document with dreal template
}
