% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{html_vignette}
\alias{html_vignette}
\title{Convert to an HTML vignette with dreal template}
\usage{
html_vignette(..., offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[rmarkdown]{html_vignette}}}

\item{offline}{Logical. TRUE if no access to internet}
}
\description{
Convert to an HTML vignette with dreal template
}
