% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{my_icon}
\alias{my_icon}
\title{Add Icon in the text}
\usage{
my_icon(name = drealdown::my_icon_list, size = 1, color = "#5a1e82")
}
\arguments{
\item{name}{icon name}

\item{size}{icon size}

\item{color}{icon color}
}
\value{
HTML object with an icon
}
\description{
Add Icon in the text
}
\examples{
my_icon("la-jsfiddle")
my_icon("drl-explosion", size = 2)
my_icon("drl-cadastre", size = 7, color = 'red')
}
