% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_icon_list.R
\docType{data}
\name{my_icon_list}
\alias{my_icon_list}
\title{Liste des icones}
\format{An object of class \code{character} of length 686.}
\usage{
my_icon_list
}
\description{
Liste des icones
}
\details{
\itemize{
\item \code{dr_*}: icones personnalisees DREAL
\item \code{la_*} : icones de la librairie \href{https://icones8.fr/line-awesome}{Line Awesome}
}
}
\keyword{data}
