% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gitbook}
\alias{gitbook}
\title{The GitBook output format with dreal template}
\usage{
gitbook(..., offline = FALSE)
}
\arguments{
\item{...}{parameters of \code{\link[bookdown]{gitbook}}}

\item{offline}{Logical. TRUE if no access to internet}
}
\description{
The GitBook output format with dreal template
}
